use size_t instead of unsigned long

+++ mmap/mmap.h
@@ -65,7 +65,7 @@ typedef struct {
      * gamma: the size of the universe that we will zero-test things at
      */
     void (*const init)(mmap_sk *const sk, size_t lambda, size_t kappa,
-                       size_t gamma, int *pows, unsigned long ncores,
+                       size_t gamma, int *pows, size_t ncores,
                        aes_randstate_t rng, bool verbose);
     void (*const clear)(mmap_sk *const sk);
     void (*const fread)(mmap_sk *const sk, FILE *const fp);
