+++ mmap/mmap_clt.c
@@ -26,7 +26,7 @@ static const mmap_pp_vtable clt_pp_vtable =
   };
 
 static void clt_state_init_wrapper(mmap_sk *const sk, size_t lambda, size_t kappa,
-                                   size_t gamma, int *pows, unsigned long ncores,
+                                   size_t gamma, int *pows, size_t ncores,
                                    aes_randstate_t rng, bool verbose)
 {
     bool new_pows = false;
